server/requirements.txt
appdirs==1.4.4
bidict==0.21.2
certifi==2021.5.30
click==8.0.1
dataclasses==0.8
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.6.0
importlib-resources==5.1.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
SQLAlchemy==1.4.19
typing-extensions==3.10.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
Werkzeug==2.0.1
zipp==3.4.1
